@@ -81,12 +85,20 @@ ext.buildInfo = { ->
     }
 }
 
-sourceCompatibility = JavaVersion.VERSION_11
-targetCompatibility = JavaVersion.VERSION_11
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
 group   = 'com.adaptris'
 version = releaseVersion
 def versionDir = "$buildDir/version"
 lombok.disableConfig = true
+
+sonar {
+  properties {
+    property "sonar.projectKey", "adaptris_interlok-solace"
+    property "sonar.organization", "adaptris-org"
+    property "sonar.host.url", "https://sonarcloud.io"
+  }
+}
 
 // Disable gradle module generation since we probably don't want
 // xxx.module files in your repository.
